[Justin][update]fix

+++ ox_common/lib/utils/ox_userinfo_manager.dart
@@ -60,7 +60,7 @@ class OXUserInfoManager {
 
   bool canVibrate = true;
   bool canSound = true;
-  double defaultZapAmount = 0.0;
+  int defaultZapAmount = 0;
 
   Future initDB(String pubkey) async {
     AppInitializationManager.shared.shouldShowInitializationLoading = true;
